# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BaseResponse {
  success: boolean;
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
  /** @nullable */
  device_id?: string | null;
  /** @nullable */
  device_model?: string | null;
  /** @nullable */
  device_os?: string | null;
}

export interface UserData {
  id: number;
  username: string;
  /** @nullable */
  full_name?: string | null;
  data_limit_mb: number;
  /** @nullable */
  expiry_date?: string | null;
}

/**
 * @nullable
 */
export type AuthResultData = {
  token?: string;
  user?: UserData;
} | null;

export interface AuthResult {
  success: boolean;
  message: string;
  banned: boolean;
  /** @nullable */
  data?: AuthResultData;
}

export interface StatsResponse {
  success: boolean;
  users: number;
  companies: number;
  servers: number;
  spoof_urls: number;
  sessions: number;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  full_name?: string | null;
  role: string;
  status: string;
  data_limit_mb: number;
  /** @nullable */
  expiry_date?: string | null;
}

export interface UserInput {
  username: string;
  password: string;
  /** @nullable */
  full_name?: string | null;
  role?: string;
  data_limit_mb?: number;
  /** @nullable */
  expiry_date?: string | null;
}

export interface UserUpdate {
  /** @nullable */
  full_name?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  data_limit_mb?: number | null;
  /** @nullable */
  expiry_date?: string | null;
}

export interface UserResponse {
  success: boolean;
  message?: string;
  data?: User;
}

export interface UsersResponse {
  success: boolean;
  data: User[];
}

export interface Company {
  id: number;
  name: string;
  name_ar: string;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  prefixes?: string | null;
  is_active: number;
}

export interface CompanyInput {
  name: string;
  name_ar: string;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  prefixes?: string | null;
}

export interface CompanyResponse {
  success: boolean;
  message?: string;
  data?: Company;
}

export interface CompaniesResponse {
  success: boolean;
  data: Company[];
}

export interface Server {
  id: number;
  company_id: number;
  server_number?: number;
  display_name: string;
  host: string;
  port: number;
  protocol: string;
  server_type: string;
  sni_hostname: string;
  payload: string;
  /** @nullable */
  company_name?: string | null;
  is_active?: number;
}

export interface ServerInput {
  company_id: number;
  server_number?: number;
  display_name: string;
  host: string;
  port?: number;
  protocol?: string;
  server_type?: string;
  sni_hostname: string;
  payload: string;
}

export interface ServerResponse {
  success: boolean;
  message?: string;
  data?: Server;
}

export interface ServersResponse {
  success: boolean;
  companies: Company[];
  servers: Server[];
  user?: UserData;
}

export interface SpoofUrl {
  id: number;
  company_id: number;
  url: string;
  /** @nullable */
  description?: string | null;
  is_active: number;
  /** @nullable */
  company_name?: string | null;
}

export interface SpoofUrlInput {
  company_id: number;
  url: string;
  /** @nullable */
  description?: string | null;
}

export interface SpoofUrlResponse {
  success: boolean;
  message?: string;
  data?: SpoofUrl;
}

export interface SpoofUrlsResponse {
  success: boolean;
  data: SpoofUrl[];
}