# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthResultData } from './authResultData';

export interface AuthResult {
  success: boolean;
  message: string;
  banned: boolean;
  /** @nullable */
  data?: AuthResultData;
}