# lib/api-zod/src/generated/types/baseResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface BaseResponse {
  success: boolean;
  message: string;
}