# lib/api-zod/src/generated/types/companiesResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from './company';

export interface CompaniesResponse {
  success: boolean;
  data: Company[];
}