# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  name_ar: string;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  prefixes?: string | null;
  is_active: number;
}