# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Server {
  id: number;
  company_id: number;
  server_number?: number;
  display_name: string;
  host: string;
  port: number;
  protocol: string;
  server_type: string;
  sni_hostname: string;
  payload: string;
  /** @nullable */
  company_name?: string | null;
  is_active?: number;
}