# lib/api-zod/src/generated/types/serverResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Server } from './server';

export interface ServerResponse {
  success: boolean;
  message?: string;
  data?: Server;
}