# lib/api-zod/src/generated/types/serversResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from './company';
import type { Server } from './server';
import type { UserData } from './userData';

export interface ServersResponse {
  success: boolean;
  companies: Company[];
  servers: Server[];
  user?: UserData;
}