# lib/api-zod/src/generated/types/spoofUrlResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { SpoofUrl } from './spoofUrl';

export interface SpoofUrlResponse {
  success: boolean;
  message?: string;
  data?: SpoofUrl;
}