# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsResponse {
  success: boolean;
  users: number;
  companies: number;
  servers: number;
  spoof_urls: number;
  sessions: number;
}