# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  username: string;
  password: string;
  /** @nullable */
  full_name?: string | null;
  role?: string;
  data_limit_mb?: number;
  /** @nullable */
  expiry_date?: string | null;
}