# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface UserResponse {
  success: boolean;
  message?: string;
  data?: User;
}