# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BTC VPN Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @nullable */
  full_name?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  data_limit_mb?: number | null;
  /** @nullable */
  expiry_date?: string | null;
}